Settings의 상수 값들을 각 관련 타입으로 이전
